requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.1
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.6.0
httpie==0.9.9
idna==2.6
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
psycopg2==2.6.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
whitenoise==3.2